.bash/prompt: use root prompt for all UID 0
The UID variable is a read-only variable in Bash, so it is safe to rely on. I might
switch to EUID if that proves to be more accurate.

As an aside, `if ((UID))` is the same as `if (($UID))`, but `if ((#))` does not work, whereas
`if (($#))` does. That is why I prefer to use the `$` inside `((` as well.