draw-ruler: initial commit
Draw a ruler across the screen. By default, it displays the column number of
the 1st, 11th, 21st, … column, and a marker on every 5 in between. You can also
specify a custom string to draw as the first parameter. Every "x" in the string
will be replaced by the column number.