git-tilde-log: always use colour
Because we explicitly use colour in the format string, we might as well
use "git log --color" to have a coloured tree (with "git ll"), too. If
we do not specify "--color", Git notices that it is being piped (through
"sed" in our case) and does not print any colour by default.