.bash/prompt: make PS4 useful for debugging
It occurred to me that you can use "dynamic" variables inside PS1 that are
not expanded until the prompt is displayed, and wondered whether you could
do the same with PS4. Turns out you can. This commit greatly enhances the
trace Bash gives you when you use "set -x" (or "bash -x"). It now shows
the name of the function and/or file, as well its line number.

I was having a lot of trouble with a shell script for a customer, and this
enabled me to find (and fix) the problem. Yay.

Of course, turns out other people have been using this same technique (in
different forms) for years:

http://raz.cx/blog/2005/08/handy-bash-debugging-trick.html
http://bashdb.sourceforge.net/bashdb.html#Interactive-Line-Tracing-Session